feat(arena-server): record logs from agents
