+++ arenaserver/container/orchestrator.go
@@ -1,14 +1,12 @@
 package container
 
 import (
-	"bufio"
 	"context"
 	"errors"
 	"io/ioutil"
 	"log"
 	"strconv"
 
-	"github.com/bytearena/bytearena/common/utils"
 	"github.com/docker/distribution/reference"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
@@ -40,32 +38,6 @@ func (orch *ContainerOrchestrator) Wait(ctner AgentContainer) error {
 		ctner.containerid.String(),
 		container.WaitConditionRemoved,
 	)
-	return nil
-}
-
-func (orch *ContainerOrchestrator) LogsToStdOut(container AgentContainer) error {
-	go func(orch *ContainerOrchestrator, container AgentContainer) {
-		reader, err := orch.cli.ContainerLogs(orch.ctx, container.containerid.String(), types.ContainerLogsOptions{
-			ShowStdout: true,
-			ShowStderr: true,
-			Follow:     true,
-			Details:    false,
-			Timestamps: false,
-		})
-
-		utils.Check(err, "Could not read container logs for "+container.AgentId.String()+"; container="+container.containerid.String())
-
-		defer reader.Close()
-
-		r := bufio.NewReader(reader)
-		scanner := bufio.NewScanner(r)
-		for scanner.Scan() {
-			text := scanner.Text()
-			log.Println(chalk.Green, container.AgentId, chalk.Reset, text)
-		}
-
-	}(orch, container)
-
 	return nil
 }
